Refactor datasets.
