+++ hyperts/datasets/__init__.py
@@ -1,0 +1,11 @@
+from hyperts.datasets.base import (
+    load_network_traffic,
+    load_random_univariate_forecast_dataset,
+    load_random_multivariate_forecast_dataset,
+)
+
+__all__ = [
+    'load_network_traffic',
+    'load_random_univariate_forecast_dataset',
+    'load_random_multivariate_forecast_dataset'
+]
